+++ lib/plugins/index.js
@@ -4,7 +4,7 @@ module.exports = [
   require('./interactiveCli'),
   require('./config.js'),
   require('./create/create.js'),
-  require('./install/install.js'),
+  require('./install.js'),
   require('./package/package.js'),
   require('./deploy.js'),
   require('./invoke/invoke.js'),
