+++ lib/plugins/install.js
@@ -2,7 +2,7 @@
 
 const BbPromise = require('bluebird');
 
-const download = require('../../utils/downloadTemplateFromRepo');
+const download = require('../utils/downloadTemplateFromRepo');
 
 class Install {
   constructor(serverless, options) {
